# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BookStore Swagger Documents!",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "노르웨이의 숲"
          },
          "description": {
            "type": "string",
            "example": "좋은 소설"
          },
          "price": {
            "type": "integer",
            "example": 15000
          },
          "stock_quantity": {
            "type": "integer",
            "example": 20
          },
          "seller_id": {
            "type": "integer",
            "example": 11
          },
          "created_at": {
            "type": "string",
            "example": "2025-12-13T17:46:41.482Z"
          }
        }
      },
      "BooksPagedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "example": "2025-12-13T17:46:41.482Z"
              },
              "page": {
                "type": "integer",
                "example": 1
              },
              "perPage": {
                "type": "integer",
                "example": 10
              },
              "sort": {
                "type": "string",
                "example": "created_at_desc"
              },
              "totalItems": {
                "type": "integer",
                "example": 123
              },
              "totalPages": {
                "type": "integer",
                "example": 13
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/books": {
      "post": {
        "summary": "책 등록",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "price",
                  "seller_id"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "책 등록 성공"
          },
          "400": {
            "description": "입력 body 누락"
          },
          "403": {
            "description": "seller 가 아닌 유저"
          }
        }
      },
      "get": {
        "summary": "책 목록 조회",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "요청할 페이지 번호 (1-indexed)",
            "required": false
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer"
            },
            "description": "페이지당 아이템 개수 (최대 100)",
            "required": false
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "정렬 기준 (price_asc, price_desc, title_asc, title_desc, created_at_asc, created_at_desc)",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "성공 (페이징된 책 배열 반환)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooksPagedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "책 상세 조회",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "책 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "404": {
            "description": "책을 찾을 수 없음"
          }
        }
      },
      "patch": {
        "summary": "책 수정",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "책 ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 성공"
          },
          "401": {
            "description": "미인증"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "id = {id} 인 책 찾을 수 없음"
          }
        }
      },
      "delete": {
        "summary": "책 삭제",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "책 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 성공"
          },
          "401": {
            "description": "미인증"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "id = {id} 인 책 찾을 수 없음"
          }
        }
      }
    },
    "/books/{bookId}/categories/{categoryId}": {
      "post": {
        "summary": "도서에 카테고리 추가",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bookId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "책 ID"
          },
          {
            "in": "path",
            "name": "categoryId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "카테고리 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "추가 성공"
          },
          "401": {
            "description": "미인증"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "id = {id} 인 책 찾을 수 없음"
          }
        }
      },
      "delete": {
        "summary": "도서에서 카테고리 제거",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bookId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "책 ID"
          },
          {
            "in": "path",
            "name": "categoryId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "카테고리 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "제거 성공"
          },
          "401": {
            "description": "미인증"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "id = {id} 인 책 찾을 수 없음"
          }
        }
      }
    },
    "/books/{bookId}/authors/{authorId}": {
      "post": {
        "summary": "도서에 저자 추가",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bookId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "책 ID"
          },
          {
            "in": "path",
            "name": "authorId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "저자 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "추가 성공"
          },
          "401": {
            "description": "미인증"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "id = {id} 인 책 찾을 수 없음"
          }
        }
      },
      "delete": {
        "summary": "도서에서 저자 제거",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bookId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "책 ID"
          },
          {
            "in": "path",
            "name": "authorId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "저자 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "제거 성공"
          },
          "401": {
            "description": "미인증"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "id = {id} 인 책 찾을 수 없음"
          }
        }
      }
    },
    "/books/{id}/reviews": {
      "post": {
        "summary": "도서에 리뷰 등록",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "책 ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content",
                  "rating"
                ],
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "등록 성공"
          },
          "404": {
            "description": "id = {id} 인 도서를 찾을 수 없습니다."
          }
        }
      },
      "get": {
        "summary": "도서에 대한 모든 리뷰 조회",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "책 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "성공 (리뷰 배열 반환)"
          }
        }
      }
    },
...
  },
...
}